add raw pattern generator
To allow more generic user-driven configuration of CPU patterns the new
`raw_pattern()` accepts a vector of bool which CPUs should be enabled or
disabled.

This can be considered a wrapper to transform a list of bools to the
set-based internal representation to be more in line with the existing
pattern generators.