add configuration for kallisto
The configuration for the test system "kallisto" uses a alderlake
processor, i.e. there are different types of cores.

This newly added test aims to explore the behavior of these cores.
It is not finished and in its current state mainly handles the pattern
creation, i.e. the mapping of function to CPU number.